escape quotes AND backslashes. Fixes UIIN-55
When querying a joined field in a CQL expression, the whole join
expression is quoted, as are the internal components of the expression,
which therefore must be escaped. And if we want to pass a string such as
"backslash-quote" through to the CQL parser, we need to escape both the
backslash and the quote in the Javascript.