make sure you have the right resource
Holy COW that was rabbit hole I just fell down. Does a cow fit in a
rabbit hole? No, no it does not, and therein lies the problem. It's
spelled out in the comments on getDerivedStateFromProps: somehow, the
values retrieved from nextProps.resources seem to be out of sync such
that they're always one click behind the most recent item you clicked on
after the first. Some extra error checking cleans this up, but it
doesn't seem like it should be necessary in the first place.