*sigh* add eslint-disable-line for import from stripes-core
This is required because stripes-core is a peer-dependency, not a
regular-dependency, and so does not get pulled in by the testing
environment before it runs ESLint.

It's still not clear what the right solution is here: perhaps, if
we're importing from stripes-core, it really is now a regular
dependency, and should be registered as one?