Display resource-type in record list
Takes advantage of the slightly richer record that comes back from
GraphQL, which includes the following fields not included in the
stripes-connect records:

* `instanceFormat`
* `instanceType`
* `edition`
* `instanceFormatId`
* `__typename`

Of these, `__typename` seems to be a GraphQL-internal field, always
with the value `Instance`, so it's not much use to us. I don't know
what instance-formats are -- that needs domain knowledge that I
lack. `edition` might be of interest, but resource-type suffices to
make the point.

There is a problem, though: because GraphQL is only being used to
assemble a rich record, and the actual querying is still left to CQL,
we're limited in our sorting options to what the CQL provides. In the
case of resource-types, that is the ID of the type rather than its
name. As a result, clicking on the "Type" column heading does sort by
type, but it's only by happy accident that right now the three types
that exist in our system happen to have IDs that sort in the same
order as their names. In other words, sorting by resource-type _seems_
to work, but it's an illusion.