console cleanup
Copy-number is an array on the backend but exposed as a non-repeatable
field on the front-end. Thus, we need to convert the value when
displaying the form (to provide the expected data-type to the Field,
which would get an array but wants a string/number) just as we don when
saving the form (which would receive a string/number but wants an
array).