clean up props-mismatch console warning
Copy-number is an array on the backend but exposed as a non-repeatable
field on the front-end. Previously, the raw array was provided to the
field, resulting in a prop-types console warning. Rather than converting
back and forth between array and scalar values, we just operate on the
first element of the array.

Replaces #691