UIIN-821 link to requests with item-id
Because a barcode is not a required field, it is possible to create
requests on items without barcodes. Formerly, the link into the requests
app would use the barcode if present and the item-id if not. This caused
problems when the barcode was applied to the item after the request was
created. Using the item-id for all links resolves the problem.

Fixes UIIN-821