refactor `it` blocks in filter test
Previous versions contained `it` blocks in functions called by multiple
tests. This test failed in CI although it ran just fine in other
environments. This refactoring changes the test structure a bit so that
functions called by multiple tests do not contain `it` blocks.

Maybe we'll get lucky and Nightmare will be happy about that?