distinguish among multi multi-select lists (#828)
It comes down to this: Nightmare renders by default in a narrow window
which causes all the `<MultiSelect>` components' options be present
under a single overlay/portal container rather than separated by the
accordion of their parent. The previous iteration of this test expected
the options to be contained under the parent accordion, which they are
on a wide screen.