don't use optional chaining just yet
It isn't clear if we only added this with the babel-7 work @doytch added
in January or if we already had it. He mentioned adding it in his
release notes but I don't see anything specifically related to it in the
actual commits.

This may work just fine as-is, but I'm going to be cautious here.