Bug fix for UIIN-1038, separate ISB/ISSN with ', '
I fixed this by altering the formatter object in TitlesView.js that's
passed down to the MulticolumList that actually displays the ISBN/ISSN.
It now runs a simple string replace function to look for any commas in
the ISBN/ISSN string and replaces them with a comma and a trailing space.
This seemed like the simplest and easiest way to get the desired behavior
(it adds less than a line of code), and doing it in the formatting object
seemed appropriate since this is where we specify what the display of
the elements is supposed to look like.  If a comma already has a space
after it, this will put an extra space in, but that shouldn't
be a problem since A: the PO insists that users will never ever manually
enter a space after a comma during data entry, and B: HTML truncates
additional spaces during display automatically.  Note that this does not
make any changes to the string that was entered and saved-only the display
of the ISBN/ISSN string is affected.