UIIN-1038 separate identifiers with ', '
I fixed this issue by writing a very small "separator" function in the
formatting object in TitlesView.js that controls how the alternate title
blocks will be  displayed, and wrapping that function around the return
value of the function that gets the identifiers.  This seemed like the
appropriate place to put
code that governs how ISBNs/ISSNs are displayed.  Nothing about this
changes the actual value the user types in, or the value stored by the
system, as the issue is with how the identifiers display. Note that the
user can manually type in spaces after commas as well, in which case,
this change will put multiple spaces after commas.  This should not be
a problem, as the PO assures me that users will never manually type in
a space after a comma, and HTMl automatically truncates multiple spaces
anyway.

Ref: [UIIN-1038](https://issues.folio.org/browse/UIIN-1038)