UIIN-1039: Preceding and Succeeding Titles
The request here was for one specific label to not be bold. I decided
the simplest way to fix that would be to apply an element-level style to
the label instructing that specific label not to display in bold. All other
labels will still be bold. I'm pretty sure the code I'm editing is only
used to generate specific labels for preceding and succeeding titles in the
edit view.

https://issues.folio.org/browse/UIIN-1039