UIIN-907: Fast add record. Access New fast add template from Inventory
I'm able to add the button and get the fast add modal to show, but the
button does not visually match the other drop-down items.  The pluggable
component seems to have no way to tell the plug-in how it should
present itself.  I looked for a place off the menu to place this, but
the entire screen is a smart component, so I can't just place UI
elements anywhere I'd like.

In short, it works, but it looks funky, and I can't see how to
do anything about that.

refs: https://issues.folio.org/browse/UIIN-907