FOLIO-2727 change filter languages
I SWEAR THIS IS TRUE: The list of available languages is generated by
stripes-components and changed in
https://github.com/folio-org/stripes-components/pull/1319 as part of
UIIN-829. It turns out the ISO-639 languages, as of 2020-08-19, do not
include simple `Spanish` but only `Spanish; Castilian` which is sort of
non-sensical to me becuase Folio itself is offered in more than one
Spanish dialect/locale (Spanish and Latin American Spanish).

The old filters test used `English` and `Spanish` as test filters, but
since `Spanish` is no longer a legit value, the test failed. Maintenant,
nous parlons Français.

Refs FOLIO-2727