UIIN-1259: Correct data validation in statistical code settings
Ui-inventory as not passing a validator function to the
controlledVocabulary component, which was causing submission errors.
Those errors were confusing because they have very generic text, leading
to confusion about what was going wrong.  I re-wrote the validation function
to check that the statistical code was unique and hooked it up to the
component. Data is now being validated properly before being passed to the back end.

Note that there are still some bugs in this particular settings pane that have
to do with CRUD operations-I have filed separate bug reports for those issues.

refs: https://issues.folio.org/browse/UIIN-1259