UIIN-1268 error message on deleting a statistical code
This was caused by an incorrect value for the 'shouldRefresh'
option in the data manifest.  That value is supposed to be a function,
not a simple boolean.  because of this, basically anytime the ControlledVocab
tried to do any write operation, it caused an error.  Changing the value to a simple
function that returns a boolean cleared up the problem.

refs:
https://issues.folio.org/browse/UIIN-1269
https://issues.folio.org/browse/UIIN-1268