UIIN-1210: Fast add permission (Inventory)
For the first two scenarios in this story, all that was required was
to ad IfPermission checking to the fast add button.  The third scenario,
which requires that users with all inventory permissions be able to see
the fast add button, required modifying the invenotry All-permissions
to add the fast add permission.

refs: https://issues.folio.org/browse/UIIN-1210