Initial over-simplified settings for single-record import
This uses `<ControlledVocab>` to allow target profiles to be browsed,
added, edited and deleted, but only supported two of the
target-profile fields: Name and URL.

The real work here was getting the permissions defined and wrestling
Okapi to the ground to get the inserted and associated with a user. I
will be documenting the key part of the process over in the _Getting
Started_ document of mod-copycat.

The present version of this settings page will need to be extended so
that it supports the Authentication, External ID Query Map, Internal
ID Embed Path and Import Profile ID fields. This will almost certainly
mean that we need to move away from using `<ControlledVocab>` --
hopefully to some other existing framework, otherwise this is going to
involve a lot of wheel-reinvention.

Towards UICPCT-1.