Allow editing all fields of target profiles
This uses the new full-width-of-the page behaviour of Inventory
settings, so that the `<SingleRecordImport>` page can fit all six
fields into a single row of `<ControlledVocab>`. It's ugly, but
functional.

We still need to do the `<EntryManager>`-based version -- and will, in
the `<TargetProfiles>` page - but this gives us something that is in a
raw way functional for the time being.

Part of UICPCT-1.