UIIN-1385 provide missing item statuses
Some item statuses available in the mod-inventory-storage enum were not
part of the list in the UI. This would cause the UI to crash when
displaying the details record of any such item. It appears that stories
to add these missing status were filed (e.g. UIIN-756, UIIN-938) but
were not unblocked when the backend work was completed. And/or, it was
not clear what impact that backend work would have.

Fixes UIIN-1385