Make the copycat interfaces optional
When the interfaces are not present, the "Import" and "Re-import" menu
items are not made available.

This is necessary because ui-inventory is part of platform-core but
mod-copycat is not: so a mandatory dependency would prevent
platform-core from building.

Is it a problem that the package file defines user-visible permissions
(`ui-inventory.single-record-import` and
`ui-inventory.settings.single-record-import`) that have
sub-permissions provided by the now-optional mod-copycat? Happily, it
is not. I have determined experimentally that the module descriptor
derived from the ui-inventory package file can be inserted in to
Okapi's `/_/proxy/modules` even in the absence of
mod-copycat. @adamdickmeiss tells me that in this situation, the
permission is stashed but not made available to be assigned to users;
but it pops into existence if mod-copycat is subsequently added.