refactor: compare up to three numbers regardless of pattern
After some conversations with folks in library-space, the conclusion was
that catalogging practices may not follow a single standard, but there
are likely many separate standards involving numbers and strings and
that there is a general expectation that numbery-looking string will be
sorted based on the numbers they contain.

This handles that, and in a generic way (separated from enumeration)
that allows this algorithm to be used for other fields that contain
numeric values, e.g. chronology, volume, etc.