UIIN-1412 parse item.enumeration numerically if possible (#1298)
After some conversations with folks in library-space, the conclusion was
that cataloging practices may not follow a single standard, but there
are likely many separate standards involving numbers and strings and
that there is a general expectation that numbery-looking strings like 
`vol.5:no.7-12` will be sorted based on the numbers they contain and
that these numbers will be handled numerically, i.e. 2 will sort ahead of
10. 

This handles that for enumeration, volume, and copyNumber.
It pulls out up to three numbers and compares them in order. 
Numbery strings sort ahead of letter-only strings.

There are lots of tests with lots of examples. 

Fixes UIIN-1412, UIIN-1416, UIIN-1425