Comment out stub route for /copycat/profiles
Its absence only causes ugly warning messages when the tests run, but
its presence causes mysteriously failures elsewhere on the test suite.

I will continue to look into this and try to find out what is going
on; but right now, what we need it to merge this damned code.