Update settings pages to reflect copycat profile changes (#1330)
The obsolete `jobProfileId` field is no longer used; instead the two
replacement fields `createJobProfileId` and `updateJobProfileId`
fields are both provided.

Also, since `targetOptions` is an object (i.e. a set of key=value
pairs) and not a string as I had assumed, I have for them moment
removed it from the relevant pages. We will want to support it, but
doing so is not trivial, so it's better for now to have settings pages
that work rather than straight-up crashing when this field is defined.

Changes have been made both the deprecated "Single record import" page
and the new "Z39.50 target profiles" page.

All of this is related to UIIN-1401.