UIIN-1440 rely on data-export interface 4.0 (#1344)
Update the okapi dependency on `data-export` to `4.0`. No information is
available about the breaking changes but in fact the changes were made
some time ago and the interface version only bumped recently, i.e. we've
been using the updated interface for some time now, but it just hadn't
been correctly labeled as updated.

Refs UIIN-1440