Avoid error message from trying to fetch /copycat/profiles
This is fetched in ImportRecordModal.js. That component was already,
rightly, not being included when the the `copycat-imports` interface
is unavailable. We now also omit it if the user doesn't have the
necessary permission, `copycat.profiles.collection.get`.

This permission is added as a sub-permission to
`ui-inventory.single-record-import`, since it's not possible to
perform an import without fetching a copycat profile.

Fixes UIIN-1447.