UIIN-1469 correctly display an item's open-request-count
Previously, the count was based on the size of the result set from a
query without a limit clause, resulting in a maximum count of the
default limit (10). Now, only a single request is retrieved (because
only a single object is used when, e.g., copying an existing request)
and the count is instead reported from the resource's `totalRecords`
value.

Refs UIIN-1469