UIIN-1451 use contributorsNames index for efficient search
Use the `contributorsNames` index from `inventory` `10.10` in place of
`contributors =/@name "..."` for vast gains in efficiency. Even though
this index was added in a minor version bump to the interface, because
we are dropping compatibility for previous versions this is a breaking
change and hence a major-version bump here in ui-inventory.

Refs UIIN-1451, RMB-797