UIIN-1483 lock stripes-cli to ~2.1.1 (#1372)
`stripes-cli` `2.1.1` locks `stripes-webpack` to `~1.1.0`, avoiding the
`1.2.0` release that adds some helpful things to do with CSV files and
the new react JSX transform but that does not play nice with BigTest for
some developers, making tests impossible to run, and, therefore,
releases impossible to release. This allows tests to be run locally in 
order to prove they run successfully, even when the fail consistently
in CI. 

In CI, do not run `stripes test karma...` since it is known to fail. Instead,
run jest tests only. As above, the BigTest suite may still be run locally in
order to demonstrate that tests pass by copying test output to a PR.

Refs UIIN-1483, STCLI-176