STRIPES-741: Add <Pluggable> for checking copyright
This commit adds a <Pluggable> to the main Instance action menu to
enable plugins that provide copyright checking functionality to add
themselves to the menu, via the new 'copyright-permissions-checker'
plugin type

This was added to enable the CLA Permissions Check plugin
(https://github.com/PTFS-Europe/ui-plugin-cla-permissions-check) to add
itself.

Additional state was added to the component to maintain the open status
of a modal that the plugin may present, this state is passed to the
plugin, along with a function to set it.

Since such plugins will require an identifier for the instance to enable
them to check copyright status, a new function getIdentifiers was added
which, when passed an instance's data, will extract a suitable
identifier to pass to the plugin