UIIN-1980: Update locations in <ViewHoldingsRecord> after edit. (#1658)
This commit includes a lot of cleanup of location-related code
that was using local resources to manage location state. Because
the resources weren't being updated properly, the locations displayed
in the location accordion would never be updated after an edit
unless the entire page were refreshed in the browser. The pane header,
on the other hand, displayed the correct location because it was using
passed-in referenceTables to determine location name from the holdings
record. Switching to that method for the location accordion as well
allows us to eliminate a lot of unnecessary code.

Refs UIIN-1980