coerce undefined to boolean
Coerce a potentially undefined state value to boolean in order to avoid
a prop-types warning. Values in conditionals are automatically coerced,
hence this isn't necessary in the many other places where it's
evaluated. Here, however, where it's just passed as a (required) prop,
we need to coerce it to a boolean value.