Back-button no longer creates duplicate instance
Previously, after using the single record import facility, using the
Back button in the browser would go back to the page that creates the
new instance rather than the page from which this facility was
invoked, and so a duplicate import was created.

This no longer happens because the redirect from the instance-creating
page to the destination now uses `history.replace` instead of
`history.push`.

Fixes UIIN-2197.