feat: Wired up settings to ItemForm
Added useConfigurationQuery hook in the same style as other hooks in use already, and made use of this in EditItem and CreateItem to pass to ItemForm.

Ideally this could just live in ItemForm, but that component is a class component and refactoring that to functional to use a hook is out of scope for this work.