bump major to v14.0.0 for new development work (#2803)
The major version was inadvertently bumped to 14 in PR #2769 and then
reverted down to 13 in PR #2793. Although the present release was not
intended to be a major release, now that some v14.x releases have
already been published to CI we are better off just keeping that change
rather than trying to undo it, which is likely to be even more
disruptive.

Co-authored-by: Oleksandr Hladchenko <85172747+OleksandrHladchenko1@users.noreply.github.com>